+++ rust/src/libponyc/ast/source.rs
@@ -290,7 +290,7 @@ pub unsafe extern "C" fn source_open(
     (*source).len = (size + 1 as libc::c_int as libc::c_long) as usize;
     let mut read: ssize_t = fread(
         (*source).m as *mut libc::c_void,
-        ::core::mem::size_of::<libc::c_char>() as libc::c_ulong,
+        ::core::mem::size_of::<libc::c_char>(),
         size as libc::c_ulong,
         fp,
     ) as ssize_t;
@@ -376,7 +376,7 @@ static mut source_pony: pony_type_t = unsafe {
     {
         let mut init = _pony_type_t {
             id: 0 as libc::c_int as u32,
-            size: ::core::mem::size_of::<source_t>() as libc::c_ulong as u32,
+            size: ::core::mem::size_of::<source_t>() as u32,
             field_count: 0 as libc::c_int as u32,
             field_offset: 0 as libc::c_int as u32,
             instance: 0 as *const libc::c_void as *mut libc::c_void,
