+++ rust/src/libponyc/ast/stringtab.rs
@@ -770,7 +770,7 @@ pub unsafe extern "C" fn stringtab_done() {
     memset(
         &mut table as *mut strtable_t as *mut libc::c_void,
         0 as libc::c_int,
-        ::core::mem::size_of::<strtable_t>() as libc::c_ulong,
+        ::core::mem::size_of::<strtable_t>(),
     );
 }
 #[c2rust::src_loc = "120:1"]
@@ -927,7 +927,7 @@ static mut strlist_pony: pony_type_t = unsafe {
     {
         let mut init = _pony_type_t {
             id: 0 as libc::c_int as u32,
-            size: ::core::mem::size_of::<strlist_t>() as libc::c_ulong as u32,
+            size: ::core::mem::size_of::<strlist_t>() as u32,
             field_count: 0 as libc::c_int as u32,
             field_offset: 0 as libc::c_int as u32,
             instance: 0 as *const libc::c_void as *mut libc::c_void,
