+++ rust/src/libponyc/codegen/genprim.rs
@@ -6993,7 +6993,7 @@ pub unsafe extern "C" fn genprim_reachable_init(mut c: *mut compile_t, mut progr
                         entity,
                         (::core::mem::size_of::<[*mut *mut ast_t; 3]>() as libc::c_ulong)
                             .wrapping_div(
-                                ::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong,
+                                ::core::mem::size_of::<*mut *mut ast_t>(),
                             )
                             .wrapping_sub(1),
                         children.as_mut_ptr(),
