+++ rust/src/libponyc/codegen/genserialise.rs
@@ -2157,7 +2157,7 @@ pub unsafe extern "C" fn genserialise(mut c: *mut compile_t, mut t: *mut reach_t
             ast_get_children(
                 (*t).ast,
                 (::core::mem::size_of::<[*mut *mut ast_t; 3]>() as libc::c_ulong)
-                    .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+                    .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
                     .wrapping_sub(1),
                 children.as_mut_ptr(),
             );
@@ -2180,7 +2180,7 @@ pub unsafe extern "C" fn genserialise(mut c: *mut compile_t, mut t: *mut reach_t
             ast_get_children(
                 (*t).ast,
                 (::core::mem::size_of::<[*mut *mut ast_t; 3]>() as libc::c_ulong)
-                    .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+                    .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
                     .wrapping_sub(1),
                 children_0.as_mut_ptr(),
             );
